17. RAG Document/4.1 RAG Document Q&A/requirements.txt
langchain
langchain-core
langchain-groq
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
streamlit
sentence-transformers